% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol-hectare.R
\name{volHa}
\alias{volHa}
\title{Tree Volume per hectare}
\usage{
volHa(x, y)
}
\arguments{
\item{x}{volume per plot}

\item{y}{plot dimension}
}
\value{
Volume per hectare
}
\description{
Tree Volume per hectare
}
\details{
Presents Tree Volume per hectare
}
\examples{
 
volHa(x = 27.85, y = 25 * 25)

}
\author{
A. Akinnagbe, P. O. Omondiagbe & S. O. Akindele
}
