% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol-newton.R
\name{volNewton}
\alias{volNewton}
\title{Tree Volume using Newton's Formula}
\usage{
volNewton(db, dm, dt, h)
}
\arguments{
\item{db}{diameter of a tree at the base in centimeters (cm)}

\item{dm}{diameter of a tree at the middle in centimeters (cm)}

\item{h}{height of a tree in meters (m)}

\item{dh}{diameter of a tree at the top in centimeters (cm)}
}
\value{
Tree volume
}
\description{
This function estimates tree volume using Newton's formula
}
\details{
Tree Volume Estimation
}
\examples{
 
volNewton(db = 46.5, dm = 27.6, dt = 11.2, h = 38.6)

}
\author{
A. Akinnagbe, P. O. Omondiagbe & S. O. Akindele
}
