% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol-plot.R
\name{volPlot}
\alias{volPlot}
\title{Volumn Plot}
\usage{
volPlot(x)
}
\arguments{
\item{x}{dataframe or matrix of trees in a plot with diameter and height}
}
\value{
total volume per plot
}
\description{
Here
}
\details{
Calculate total tree volume per plot
}
