% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol-huber.R
\name{volHuber}
\alias{volHuber}
\title{Tree Volume using Huber's Formula}
\usage{
volHuber(dm, h)
}
\arguments{
\item{dm}{diameter of a tree at the middle in centimeters (cm)}

\item{h}{height of a tree in meters (m)}
}
\value{
Tree volume
}
\description{
This function estimates tree volume using Huber's formula
}
\details{
Tree Volume Estimation
}
\examples{
 
volHuber(dm = 35.6, h = 20)

}
\author{
A. Akinnagbe, P. O. Omondiagbe & S. O. Akindele
}
