% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol-plot.R
\name{volPlot}
\alias{volPlot}
\title{Summary of Tree Volume in a plot}
\usage{
volPlot(x, y, dbCol = NULL, dmCol = NULL, dtCol = NULL,
  hCol = NULL)
}
\arguments{
\item{x}{dataframe of trees in a plot with diameter and height}

\item{y}{function to calculate tree volume. Enter volHuber for Huber, volSmalian for Smalian and volNewton for Newton.}

\item{dbCol}{column number of the diameter at the base on the dataframe}

\item{dmCol}{column number of the diameter at the middle (otherwise known as dbh in some cases) on the dataframe}

\item{dtCol}{column number of the diameter at the top on the dataframe}

\item{hCol}{column number of the height on the dataframe}
}
\value{
Tree volume summary in a plot or stand
}
\description{
Basic statistics of tree volume in a plot
}
\details{
Presents summary of tree volume in a plot
}
\examples{
x <- data.frame(db = c(38.40, 28.10, 63.10, 64.5), dm = c(25.4, 18.6, 41.6, 42.6), dt = c(12.7, 9.3, 20.8, 21.3), h = c(38.4, 32.3, 68.3, 69.5))
volPlot(x, volHuber, dmCol = 2, hCol = 4)
volPlot(x, volSmalian, dbCol = 1, dtCol = 3, hCol = 4)
volPlot(x, volNewton, dbCol = 1, dmCol = 2, dtCol = 3, hCol = 4)

}
\author{
A. Akinnagbe, P. O. Omondiagbe & S. O. Akindele
}
