% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol-smalian.R
\name{volSmalian}
\alias{volSmalian}
\title{Tree Volume Estimation}
\usage{
volSmalian(db, dt, h)
}
\arguments{
\item{db}{diameter of a tree at the base in centimeters (cm)}

\item{h}{height of a tree in meters (m)}

\item{dm}{diameter of a tree at the middle in centimeters (cm)}

\item{dh}{diameter of a tree at the top in centimeters (cm)}
}
\value{
Tree volume
}
\description{
This function estimates tree volume using Smalian's formula
}
\author{
A. Akinnagbe, P. O. Omondiagbe & S. O. Akindele
}
