% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol-trees.R
\name{volTrees}
\alias{volTrees}
\title{Tree Volumn}
\usage{
volTrees(x)
}
\arguments{
\item{x}{dataframe or matrix of trees in a plot with diameter and height}
}
\value{
dataframe containing volumes of individual trees
}
\description{
Calculate volume of individual trees in a plot or stand
}
